+++ Shop.Bsl.Flow/DataClasses.cs
@@ -1,4 +1,5 @@
 ﻿//Adapted from Microsoft.AspNet.OData.Test.Query.Expressions
+using Microsoft.OData.Edm;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -32,11 +33,17 @@ namespace Shop.Bsl.Flow
         public DateTimeOffset DiscontinuedOffset { get; set; }
         public TimeSpan DiscontinuedSince { get; set; }
 
+        public Date DateProperty { get; set; }
+        public Date? NullableDateProperty { get; set; }
+
         public DateOnly DateOnlyProperty { get; set; }
         public DateOnly? NullableDateOnlyProperty { get; set; }
 
         public Guid GuidProperty { get; set; }
         public Guid? NullableGuidProperty { get; set; }
+
+        public TimeOfDay TimeOfDayProperty { get; set; }
+        public TimeOfDay? NullableTimeOfDayProperty { get; set; }
 
         public TimeOnly TimeOnlyProperty { get; set; }
         public TimeOnly? NullableTimeOnlyProperty { get; set; }