+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* 2026-09-02 - AB#221: Removed unused calss members form ConstructorXmlBuilder.
+* 2026-09-03 - AB#221: Fix bug in ConstructorXmlBuilder logic.
+* 2026-09-02 - AB#221: Removed unused class members form ConstructorXmlBuilder.
 * 2026-09-01 - AB#221: Removed model specific XML generation.
 * 2026-09-01 - AB#221: 245/303 tests working for Lambda Expression XML definiton.
 * 2026-08-30 - AB#221: Geneerated XML matches Contoso filters and selectors.